requirements.txt
absl-py==0.15.0
accelerate==1.2.1
addict==2.4.0
aenum==3.1.15
aesthetic-predictor-v2-5==2024.12.18.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioice==0.10.1
aiortc==1.10.1
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.6
allure-pytest==2.13.5
allure-python-commons==2.13.5
altair==4.2.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
args==0.1.0
assertpy==1.1
astroid==2.15.8
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
av==13.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
black==23.1.0
blinker==1.9.0
braceexpand==0.1.7
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
clint==0.5.1
clip==0.2.0
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.3.0
cpm-kernels==1.0.11
cryptography==43.0.3
curlify==2.2.1
cycler==0.12.1
dashscope==1.22.2
datasets==3.3.2
decorator==5.1.1
decord==0.6.0
deepspeed==0.16.2
delayed-assert==0.3.6
diffusers==0.32.2
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
easydict==1.13
einops==0.6.1
einops-exts==0.0.4
entrypoints==0.4
esda==2.5.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
exrex==0.12.0
facexlib==0.3.0
fairscale==0.4.13
Faker==33.3.0
fast-map==0.2.0
fastapi==0.115.6
fastapi-utils==0.8.0
ffmpy==0.5.0
filelock==3.17.0
filterpy==1.4.5
flake8==6.0.0
flash-attention==1.0.0
flash_attn==2.7.4.post1
Flask==3.1.0
Flask-BasicAuth==0.2.0
Flask-Cors==5.0.0
flatbuffers==1.12
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
gast==0.4.0
gdown==5.2.0
geopandas==1.0.1
gevent==21.12.0
geventhttpclient==2.3.3
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-crc32c==1.7.1
google-pasta==0.2.0
gradio==4.16.0
gradio_client==0.8.1
greenlet==1.1.2
grpcio==1.34.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.1.0
hf-xet==1.0.3
hjson==3.1.0
-e git+https://github.com/tgxs002/HPSv2.git@866735ecaae999fa714bd9edfa05aa2672669ee3#egg=hpsv2
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.30.1
icecream==2.1.4
identify==2.6.7
idna==3.10
ifaddr==0.2.0
iglovikov-helper-functions==0.0.53
image-reward==1.5
imagecorruptions==1.1.2
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
influxdb-client==1.48.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.0
isort==5.13.2
itsdangerous==2.2.0
jaxtyping==0.2.36
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jpeg4py==0.1.4
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
kornia==0.8.0
kornia_rs==0.1.8
latex2mathml==3.77.0
lazy-object-proxy==1.10.0
lazy_loader==0.4
libpysal==4.8.1
-e git+https://github.com/LLaVA-VL/LLaVA-NeXT@79ef45a6d8b89b92d7a8525f077c3a3a9894a87d#egg=llava
llvmlite==0.43.0
lmdb==1.6.2
locust==2.12.1
loguru==0.7.3
lpips==0.1.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mmcv==2.2.0
mmdet==3.3.0
mmengine==0.10.7
modelcards==0.1.6
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.33.0
networkx==3.2.1
ninja==1.11.1.3
nodeenv==1.9.1
numba==0.60.0
numpy==1.23.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
omegaconf==2.3.0
open_clip_torch==2.30.0
openai==1.78.0
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pandas==1.4.4
parso==0.8.4
pathspec==0.12.1
peft==0.14.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
pre_commit==4.1.0
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.3
psutil==5.9.8
psycogreen==1.0.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyByteBuffer==1.0.5
pycocotools==2.0.8
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pyext==0.7
pyflakes==3.0.1
Pygments==2.19.1
pyiqa==0.1.13
PyJWT==2.10.1
pylibsrtp==0.12.0
pylint==2.17.0
Pympler==1.1
pynvml==12.0.0
pyogrio==0.10.0
pyOpenSSL==24.3.0
pyparsing==3.2.1
pyproj==3.6.1
PySocks==1.7.1
pytest==7.2.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-order==1.3.0
pytest-reportportal==5.1.5
pytest-rerunfailures==15.0
pytest-split==0.8.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
reactivex==4.0.4
referencing==0.36.0
regex==2024.11.6
reportportal-client==5.2.5
requests==2.32.3
requests-oauthlib==2.0.0
retinaface==1.1.1
retinaface-pytorch==0.0.7
rich==13.9.0
roundrobin==0.0.4
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.4
safetensors==0.5.0
scenedetect==0.6.6
scikit-image==0.24.0
scikit-learn==1.2.2
scikit-video==1.1.11
scipy==1.13.1
seaborn==0.13.2
selenium==4.21.0
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.1.99
sentry-sdk==2.19.2
setproctitle==1.3.4
shapely==2.0.7
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
simsimd==6.2.1
six==1.15.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
streamlit==1.10.0
streamlit-webrtc==0.42.0
stringzilla==3.12.5
svgwrite==1.4.3
sympy==1.13.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
toolz==1.0.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
trio==0.28.0
trio-websocket==0.11.1
triton==3.1.0
trl==0.15.1
typeguard==4.4.1
typer==0.15.2
typing_extensions==4.12.2
tyro==0.7.3
tzdata==2024.2
tzlocal==5.3.1
unitest==1.4.34
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
virtualenv==20.29.2
wandb==0.19.1
watchdog==6.0.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
webdataset==0.2.100
webdriver-manager==4.0.2
websocket-client==1.4.2
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.12.1
wsproto==1.2.0
xformers==0.0.29.post1
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
